Fix most tests on python 2.7
This commit fixes the majority of the issues exposed by running the
unit tests. Most of these issues were improper imports or other typos
that occurred during the initial pivot or pep8 fixes. There are still
a few tests that fail (for what looks like real reasons) but those will
be fixed in a follow up patch. Additionally py3 has more failures due
to unicode fun, which will also be fixed in a follow up.