Fix python3 binary encoding failures
Early in the pivot to pysubunit six.binary_type() was being used to
replace the usage of testtools compat layer. However this was done
haphazardly and strings need to be encoded on python3 to treat it as a
bytes type. To avoid this for right now this commit switches back to
using the compat layer from testtools (which is just calling
.encode('latin-1') on the input with python 3) Longer term I think
removing the dependence on testtools compat layer is something we'll do.
But, at this point in the pivot to pysubunit this is just busy work.