Stop relying on iso8601 to parse dates
The iso8601 library's date_parse function was behaving differently
between py2 and py3. The iso8601 docs also recommend the use of
python-dateutil for more full featured parsing. This commit does just
that and switches to use python-dateutil for the one place we call
iso8601.date_parse. This also has the nice side effect of fixing the
last failures on python 3.6.